+++ src/touchscreengui.h
@@ -99,7 +99,7 @@ public:
 	bool isButton(const SEvent &event);
 
 	/* handle released hud buttons */
-	bool isReleaseButton(int eventID);
+	bool isReleaseButton(size_t eventID);
 
 	/* step handler */
 	void step(float dtime);
@@ -166,7 +166,7 @@ private:
 	ISimpleTextureSource*   m_texturesource;
 	v2u32                   m_screensize;
 	std::map<int,rect<s32> > m_hud_rects;
-	std::map<int,irr::EKEY_CODE> m_hud_ids;
+	std::map<size_t,irr::EKEY_CODE> m_hud_ids;
 	bool                    m_visible; // is the gui visible
 
 	/* value in degree */
@@ -189,7 +189,7 @@ private:
 	touch_gui_button_id getButtonID(s32 x, s32 y);
 
 	/* gui button by eventID */
-	touch_gui_button_id getButtonID(int eventID);
+	touch_gui_button_id getButtonID(size_t eventID);
 
 	/* check if a button has changed */
 	void handleChangedButton(const SEvent &event);
@@ -203,7 +203,7 @@ private:
 	void loadButtonTexture(button_info* btn, const char* path, rect<s32> button_rect);
 
 	struct id_status{
-		int id;
+		size_t id;
 		int X;
 		int Y;
 	};
@@ -218,13 +218,13 @@ private:
 	bool isHUDButton(const SEvent &event);
 
 	/* handle released hud buttons */
-	bool isReleaseHUDButton(int eventID);
+	bool isReleaseHUDButton(size_t eventID);
 
 	/* handle double taps */
 	bool doubleTapDetection();
 
 	/* handle release event */
-	void handleReleaseEvent(int evt_id);
+	void handleReleaseEvent(size_t evt_id);
 
 	/* get size of regular gui control button */
 	int getGuiButtonSize();
